Create sliding_mode_pendulum.jl
This Julia code simulates the dynamics of an inverted pendulum on a cart, controlled by a sliding mode controller. The system is described by a set of ordinary differential equations (ODEs), which model the motion of the pendulum and the cart. The code uses the DifferentialEquations.jl package to solve these ODEs numerically.